feat: Add WaterfallLayout to RAC, and update GridLayout
